Enable negative caching of non-existent tiddlers.
Make it so if a tiddler does not exist, we cache that result.

This makes it so we don't have to go to the datastore to find out
(a second and subsequent time) if a tiddler does not exist, instead
the cache can tell us. Because of the way invalidation works, if
anything gets written to the bag that the missing tiddler would be
in, the negative cache of the missing tiddler is flushed.

This has been done primarily for tiddlyspace where the MAGICUSER
and MAPUSER bags are checked for every request, to adjust the user
object. However for most users the target tiddler does not exist,
but we go to the database (twice!) every single request.

see for additional:
http://cdent.tiddlyspace.com/bags/cdent_public/tiddlers/20110513.html