Only cache lists if config['memcache.cache_lists'] is True
Turns out bags, recipes and user lists are sufficiently big to not
be worth caching. And, for those systems where the speed is needed
is better to load tons of data at once straight out of database,
in which case 'store' needs to be set on the entities, and that
cannot be pickled, so the storage doesn't happen anyway.

release as 0.9.11